refactor: 피드백 반영, viewmodel 패키지명 정리, 불필요한 의존성 제거
